TST: move tests to _basis.py
